edgexoi: yellow/blue instead of yellow/green
I am colorblind (I can never remember which particular variety) and I
bright yellow and green nearly impossible to distinguish. When doing
Slitherlink on penpa I am constantly double checking zoomed in
on high brightness to make sure I can tell which color is which. With
this replacement, it's easy to tell the difference. Blue is also light enough
that it should not occlude clues, unlike other potential choices.