Fix fit-to-window on some platforms
Fit-to-window and other options that rely on it (such as rotate) were
not working properly on some platforms (I can reproduce both on Firefox/Linux
and on Firefox/Android, although it may be dependent on specific hardware too).

The reason seems to be that some renderers unexpectedly produce not-perfectly-white
pixels throughout white areas, and the fit code was interpreting these as
not trimmable, resulting in a fitted area that was much larger than desired.