in akari mode, don't forbid creating bulbs on coloured cells
Clicking on a coloured empty cell on desktop in akari mode didn't
create a bulb because the code was explicitly testing for a surface
element.

I'm not sure what the intention was, but the result was an uncaught
TypeError as it fell through to a case that assumed there was a symbol
element at the cell.