Add support for special file saving for .url and .desktop.
In the course of doing this, I had to revamp some of the handling for filename checking and ended up condensing
some of the code into reusable functions, as well as adding input patterns on the input for the filename when
the user is entering their preferred filename. This also affects the filename entry for saving image exports.